3.1

+++ src/main/java/com/vitali/dto/order/OrderCreateDto.java
@@ -1,13 +1,13 @@
 package com.vitali.dto.order;
 
-import com.vitali.constants.OrderStatus;
+import com.vitali.entities.enums.OrderStatus;
 import lombok.Builder;
 import lombok.Value;
 
 @Builder
 @Value
 public class OrderCreateDto {
-    Long cartId;
+    Integer cartId;
 //    LocalDateTime createTime;
     String inform;
     OrderStatus orderStatus;
